+++ bindgen/src/main/java/org/bindgen/processor/generators/FieldPropertyGenerator.java
@@ -49,6 +49,7 @@ public class FieldPropertyGenerator implements PropertyGenerator {
 		this.addInnerClassSet();
 		this.addInnerClassSetWithRoot();
 		this.addInnerClassGetContainedTypeIfNeeded();
+		this.addInnerClassSerialVersionUID();
 	}
 
 	private void addOuterClassBindingField() {
@@ -121,6 +122,7 @@ public class FieldPropertyGenerator implements PropertyGenerator {
 
 	private void addInnerClassSet() {
 		GMethod set = this.innerClass.getMethod("set").argument(this.property.getSetType(), this.property.getName());
+		set.addAnnotation("@Override");
 		if (this.isFinal) {
 			set.body.line("throw new RuntimeException(this.getName() + \" is read only\");");
 			return;
@@ -133,6 +135,7 @@ public class FieldPropertyGenerator implements PropertyGenerator {
 
 	private void addInnerClassSetWithRoot() {
 		GMethod setWithRoot = this.innerClass.getMethod("setWithRoot(R root, {} {})", this.property.getSetType(), this.property.getName());
+		setWithRoot.addAnnotation("@Override");
 		if (this.isFinal) {
 			setWithRoot.body.line("throw new RuntimeException(this.getName() + \" is read only\");");
 			return;
@@ -150,6 +153,10 @@ public class FieldPropertyGenerator implements PropertyGenerator {
 			GMethod getContainedType = this.innerClass.getMethod("getContainedType").returnType("Class<?>").addAnnotation("@Override");
 			getContainedType.body.line("return {};", this.property.getContainedType());
 		}
+	}
+
+	private void addInnerClassSerialVersionUID() {
+		this.innerClass.getField("serialVersionUID").type("long").setStatic().setFinal().initialValue("1L");
 	}
 
 	public TypeElement getPropertyTypeElement() {