+++ docs/changelog.md
@@ -6,10 +6,17 @@ title: Changelog
 Changelog
 =========
 
+## 2.6 - 2009-11-26
+
+* Fix: Protected bindings are now `set`-able
+* Fix: Generated binding inner classes have `@Override` on `set` and `setWithRoot` methods
+* Fix: Generated binding classes have a `serialVersionUID=1L` to avoid compiler warnings
+* Changed: Moved `org/bindgen.gwt.xml` to `org/bindgen/Bindgen.gwt.xml`
+
 ## 2.5 - 2009-11-22
 
-* Changed: Binding extends `Serializable`
-* Changed: `GenericObjectBindingPath` is parameterized on its bound type `T`
+* Changed: Binding extends `Serializable` (Igor Vaynberg)
+* Changed: `GenericObjectBindingPath` is parameterized on its bound type `T` (Igor Vaynberg)
 
 ## 2.4 - 2009-11-19